requirements.txt
requests
beautifulsoup4
pandas
gspread
google-auth
sqlalchemy
psycopg2-binary
pytest
pytest-cov
python-dotenv